fixed bug to include t2 phase

+++ bids/rosmap_addprotocol.py
@@ -12,18 +12,20 @@ check_every = 1000
 if __name__ == "__main__":
     
     os.chdir(bids_dir)
-    datlog = pandas.read_csv(datlog_pth,index_col-0)
+    datlog = pandas.read_csv(datlog_pth,index_col=0)
 
     print('Adding ROSMAP protocol to acq- part of BIDS filename')
+    count = 0
     for i,row in datlog.iterrows():
         if count % check_every == 0:
-            print('working on %s of %s'%((count+1),len(ds)))
+            print('working on %s of %s'%((count+1),len(datlog)))
         pth = row.new_path
         prot = row.ScannerGroup + row.Protocol
         acq = pth.split('acq-')[-1].split('_')[0]
         npth = pth.replace('acq-%s'%acq,'acq-%s%s'%(prot,acq))
         datlog.loc[i,'new_path'] = npth
-        if row.ext == 'json' and row.Category == 'fmap':
+        if row.ext == 'json' and (row.Category == 'fmap' or row.Modality=='phase'):
+            if not os.path.exists(pth): continue
             with open(pth) as json_data:
                 j = json.load(json_data)
             if 'IntendedFor' in j.keys():
@@ -35,7 +37,8 @@ if __name__ == "__main__":
                     with open(pth, 'w') as fp:
                        json.dump(j, fp,sort_keys=True, indent=4)
                     datlog.loc[i,'intendedfor'] = inpth
-        os.rename(pth,npth)
+        if os.path.exists(pth):    
+            os.rename(pth,npth)
         count+=1
 
     datlog.to_csv(datlog_pth,index=False)
